Move snapshot things into individual files.
