+++ include/edyn/networking/packet/create_entity.hpp
@@ -2,7 +2,7 @@
 #define EDYN_NETWORKING_PACKET_CREATE_ENTITY_HPP
 
 #include <vector>
-#include "edyn/networking/util/pool_snapshot.hpp"
+#include "edyn/networking/util/registry_snapshot.hpp"
 
 namespace edyn::packet {
