+++ include/edyn/networking/packet/entity_response.hpp
@@ -2,7 +2,7 @@
 #define EDYN_NETWORKING_PACKET_ENTITY_RESPONSE_HPP
 
 #include <vector>
-#include "edyn/networking/util/pool_snapshot.hpp"
+#include "edyn/networking/util/registry_snapshot.hpp"
 
 namespace edyn::packet {
